Fixes warning about int size_t mixing

+++ include/nitro/broken_options/options.hpp
@@ -93,7 +93,8 @@ namespace broken_options
         {
             if (i < 0)
             {
-                i += positionals_.size();
+                // if you ever manage to have more than 2^31 positionals, I owe you a beer.
+                i += static_cast<int>(positionals_.size());
             }
             return positionals_.at(i);
         }
